# Fireblocks.swagger.yml
...
openapi: 3.0.0
info:
  title: Fireblocks API
  version: "1.5.0"
servers:
  - url: 'https://api.fireblocks.io/v1'
paths:
...
  /transactions:
    get:
      summary: Gets a list of transactions under the tenant
      parameters:
        - in: query
          name: before
          description: >-
            Unix timestamp in milliseconds. Returns only transactions created before the
            specified date
          required: false
          schema:
            type: string
        - in: query
          name: after
          description: >-
            Unix timestamp in milliseconds. Returns only transactions created after the
            specified date
          required: false
          schema:
            type: string
        - in: query
          name: status
          description: >-
            Comma-separated list of statuses. Returns only transactions with
            one of the specified statuses
          required: false
          schema:
            type: string
        - in: query
          name: orderBy
          description: The field to order the results by
          required: false
          schema:
            type: string
            enum:
              - createdAt
              - lastUpdated
        - in: query
          name: limit
          description: >-
            Limits the number of results. If not provided, a limit of 200 will be used. The maximum allowed limit is 500
          required: false
          schema:
            type: integer
            minimum: 1
            default: 200
        - in: query
          name: sourceType
          description: The source type of the transaction
          required: false
          schema:
            type: string
            enum:
              - VAULT_ACCOUNT
              - EXCHANGE_ACCOUNT
              - INTERNAL_WALLET
              - EXTERNAL_WALLET
              - FIAT_ACCOUNT
              - NETWORK_CONNECTION
              - COMPOUND
              - UNKNOWN
              - GAS_STATION
              - OEC_PARTNER
        - in: query
          name: sourceId
          description: The source ID of the transaction
          required: false
          schema:
            type: string
        - in: query
          name: destType
          description: The destination type of the transaction
          required: false
          schema:
            type: string
            enum:
              - VAULT_ACCOUNT
              - EXCHANGE_ACCOUNT
              - INTERNAL_WALLET
              - EXTERNAL_WALLET
              - FIAT_ACCOUNT
              - NETWORK_CONNECTION
              - COMPOUND
              - ONE_TIME_ADDRESS
              - OEC_PARTNER
        - in: query
          name: destId
          description: The destination ID of the transaction
          required: false
          schema:
            type: string
        - in: query
          name: assets
          description: A list of assets to filter by, seperated by commas
          required: false
          schema:
            type: string
        - in: query
          name: txHash
          description: Returns only results with a specified txHash
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of transactions
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionReferenceResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Creates a new transaction
      responses:
        '200':
          description: A transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
        default:
          $ref: '#/components/responses/Error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
  '/transactions/estimate_fee':
    post:
      summary: >-
        Estimates the transaction fee for a given transaction request
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Estimated fees respose
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EstimatedTransactionFeeResponse'
        default:
          $ref: '#/components/responses/Error'
  '/transactions/{txId}':
    get:
      summary: Returns a transaction by ID
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction to return
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: An Transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        default:
          $ref: '#/components/responses/Error'
  '/transactions/external_tx_id/{externalTxId}/':
    get:
      summary: Returns transaction by external transaction ID
      parameters:
        - in: path
          name: externalTxId
          required: true
          description: The external ID of the transaction to return
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: An Transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        default:
          $ref: '#/components/responses/Error'
...
  '/transactions/{txId}/drop':
    post:
      summary: Drop ETH based transaction by ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DropTransactionRequest'
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: Created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DropTransactionResponse'
        default:
          $ref: '#/components/responses/Error'
  '/transactions/{txId}/cancel':
    post:
      summary: Cancels a transaction by ID
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction to cancel
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: An Transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CancelTransactionResponse'
        default:
          $ref: '#/components/responses/Error'
  '/transactions/{txId}/freeze':
    post:
      summary: Freezes a transaction by ID
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction to freeze
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: freeze response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FreezeTransactionResponse'
  '/transactions/{txId}/unfreeze':
    post:
      summary: Unfreezes a transaction by ID
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction to unfreeze
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: Unfreeze response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnfreezeTransactionResponse'
  '/transactions/validate_address/{assetId}/{address}':
    get:
      summary: Check if given address is valid
      parameters:
        - in: path
          name: assetId
          required: true
          description: The asset of the address
          schema:
            type: string
        - in: path
          name: address
          required: true
          description: The address to validate
          schema:
            type: string
      responses:
        '200':
          description: An Transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ValidateAddressResponse'
        default:
          $ref: '#/components/responses/Error'
  # TXHASH
...
security:
  - bearerTokenAuth: [ ]
  - ApiKeyAuth: [ ]
components:
  responses:
    Error:
      description: Error Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    TransferPeerPath:
      type: object
      properties:
        type:
          type: string
        id:
          type: string
          nullable: true
      required:
        - type
    DestinationTransferPeerPath:
      allOf:
        - $ref: '#/components/schemas/TransferPeerPath'
        - type: object
          properties:
            oneTimeAddress:
              $ref: '#/components/schemas/OneTimeAddress'
    CreateTransactionResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
    CancelTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
    UnfreezeTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
    FreezeTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
...
    TransactionReferenceResponse:
      type: object
      properties:
        id:
          type: string
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
          nullable: true
        source:
          $ref: '#/components/schemas/TransferPeerPathResponse'
          nullable: true
        destination:
          $ref: '#/components/schemas/TransferPeerPathResponse'
          nullable: true
        requestedAmount:
          description: The amount requested by the user
          type: string
          nullable: true
        amount:
          description: If the transfer is a withdrawal from an exchange, the actual amount that was requested to be transferred. Otherwise, the requested amount
          type: string
          nullable: true
        netAmount:
          description: The net amount of the transaction, after fee deduction
          type: string
          nullable: true
        amountUSD:
          description: The USD value of the requested amount
          type: string
          nullable: true
        serviceFee:
          description: The total fee deducted by the exchange from the actual requested amount (serviceFee = amount - netAmount)
          type: string
          nullable: true
        networkFee:
          description: The fee paid to the network
          type: string
          nullable: true
        createdAt:
          type: number
          format: decimal
          description: Unix timestamp
          nullable: true
        lastUpdated:
          type: number
          format: decimal
          description: Unix timestamp
        status:
          type: string
        txHash:
          type: string
        tag:
          type: string
        subStatus:
          $ref: '#/components/schemas/TransactionSubStatus'
          nullable: true
        destinationAddress:
          type: string
        sourceAddress:
          type: string
        destinationAddressDescription:
          type: string
        destinationTag:
          type: string
        signedBy:
          type: array
          items:
            type: string
        createdBy:
          type: string
        rejectedBy:
          type: string
        addressType:
          type: string
        note:
          type: string
        exchangeTxId:
          type: string
        feeCurrency:
          type: string
        operation:
          $ref: '#/components/schemas/TransactionOperation'
        networkRecords:
          type: array
          items:
            $ref: '#/components/schemas/NetworkRecord'
        amlScreeningResult:
          $ref: '#/components/schemas/AmlScreeningResult'
        customerRefId:
          type: string
        numOfConfirmations:
          type: number
          format: decimal
          nullable: true
        signedMessages:
          type: array
          items:
            $ref: '#/components/schemas/SignedMessage'
        extraParameters:
          type: object
          properties: {}
        externalTxId:
          type: string
        replacedTxHash:
          type: string
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponseDestination'
        blockInfo:
          $ref: '#/components/schemas/BlockInfo'
        authorizationInfo:
          $ref: '#/components/schemas/AuthorizationInfo'
        amountInfo:
          $ref: '#/components/schemas/AmountInfo'
        index:
          type: number
          format: decimal
          nullable: true
        blockchainIndex:
          type: string
...
    EstimatedTransactionFeeResponse:
      type: object
      properties:
        low:
          $ref: '#/components/schemas/TransactionFee'
        medium:
          $ref: '#/components/schemas/TransactionFee'
        high:
          $ref: '#/components/schemas/TransactionFee'
      required:
        - low
        - medium
        - high
...
    TransactionRequest:
      type: object
      properties:
        assetId:
          type: string
        source:
          $ref: '#/components/schemas/TransferPeerPath'
        destination:
          $ref: '#/components/schemas/DestinationTransferPeerPath'
        amount:
          type: string
        fee:
          type: string
          nullable: true
          description: >-
            - For BTC based assets, the fee per bytes in the asset's smallest unit (Satoshi/Latoshi etc.)

            - For XRP, the fee for the transaction
          oneOf:
            - type: number
            - type: string
        feeLevel:
          nullable: true
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
        priorityFee:
          nullable: true
          description: >-
            - For ETH based assets only, the fee for eip-1559 transaction pricing mechanism. Value is in gwei.
          oneOf:
            - type: number
            - type: string
        failOnLowFee:
          nullable: true
          type: boolean
        maxFee:
          nullable: true
          description: >-
            - For ETH based assets only, must be provided when using the priorityFee parameter. This will be used to limit the fee max possible fee (according to eip-1559). Value is in gwei.

            - For other assets, it fails a transaction when the automatically selected fee is higher than the provided value.
          type: string
        gasPrice:
          nullable: true
          description: >-
            For ETH based asset only. this will be used instead of the fee property. Value is in gwei
          oneOf:
            - type: number
              format: decimal
            - type: string
        gasLimit:
          nullable: true
          oneOf:
            - type: number
              format: decimal
            - type: string
        networkFee:
          nullable: true
          oneOf:
            - type: number
            - type: string
        note:
          type: string
        autoStaking:
          nullable: true
          type: boolean
        networkStaking:
          nullable: true
          oneOf:
            - type: number
            - type: string
        cpuStaking:
          nullable: true
          oneOf:
            - type: number
            - type: string
        extraParameters:
          type: object
          properties: {}
        operation:
          $ref: '#/components/schemas/TransactionOperation'
        customerRefId:
          type: string
        replaceTxByHash:
          type: string
        externalTxId:
          type: string
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/TransactionRequestDestination'
        treatAsGrossAmount:
          nullable: true
          type: boolean
        forceSweep:
          nullable: true
          type: boolean
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: number
          format: decimal
...
    DropTransactionRequest:
      type: object
      properties:
        txId:
          type: string
        feeLevel:
          type: string
        gasPrice:
          type: string
    DropTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
        transactions:
          type: array
          items:
            type: string
...
    ValidateAddressResponse:
      type: object
      properties:
        isValid:
          type: boolean
        isActive:
          type: boolean
        requiresTag:
          type: boolean
...